+++ SpaceEconSim/WindowManager.cpp
@@ -27,7 +27,7 @@ bool WindowManager::CheckQuitNextUpdate()
 	return m_IsQuittingNextUpdate;
 }
 
-bool WindowManager::Display(float a_Dt)
+bool WindowManager::Display(double a_Dt)
 {
 	//show next frame
 	m_pSFMLRenderWindow->display();
@@ -86,7 +86,7 @@ void WindowManager::CleanUp()
 sf::Vector2f WindowManager::GetWindowDim()
 {
 	sf::Vector2f out;
-	out.y = (float)m_pSFMLRenderWindow->getSize().y;
-	out.x = (float)m_pSFMLRenderWindow->getSize().x;
+	out.y = (double)m_pSFMLRenderWindow->getSize().y;
+	out.x = (double)m_pSFMLRenderWindow->getSize().x;
 	return out;
 }
